move RemoteValue and child classes to new remote_value module
