+++ test/remote_value_tests/remote_value_1count_test.py
@@ -1,15 +1,15 @@
-"""Unit test for RemoteValueSceneNumber objects."""
+"""Unit test for RemoteValue1Count objects."""
 import asyncio
 import unittest
 
 from xknx import XKNX
-from xknx.devices import RemoteValueSceneNumber
-from xknx.exceptions import ConversionError, CouldNotParseTelegram
+from xknx.exceptions import CouldNotParseTelegram
 from xknx.knx import DPTArray, DPTBinary, GroupAddress, Telegram
+from xknx.remote_value import RemoteValue1Count
 
 
-class TestRemoteValueSceneNumber(unittest.TestCase):
-    """Test class for RemoteValueSceneNumber objects."""
+class TestRemoteValue1Count(unittest.TestCase):
+    """Test class for RemoteValue1Count objects."""
 
     def setUp(self):
         """Set up test class."""
@@ -23,63 +23,54 @@ class TestRemoteValueSceneNumber(unittest.TestCase):
     def test_to_knx(self):
         """Test to_knx function with normal operation."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueSceneNumber(xknx)
-        self.assertEqual(remote_value.to_knx(11), DPTArray((0x0A, )))
+        remote_value = RemoteValue1Count(xknx)
+        self.assertEqual(remote_value.to_knx(100), DPTArray((0x64, )))
 
     def test_from_knx(self):
         """Test from_knx function with normal operation."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueSceneNumber(xknx)
-        self.assertEqual(remote_value.from_knx(DPTArray((0x0A, ))), 11)
-
-    def test_to_knx_error(self):
-        """Test to_knx function with wrong parametern."""
-        xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueSceneNumber(xknx)
-        with self.assertRaises(ConversionError):
-            remote_value.to_knx(100)
-        with self.assertRaises(ConversionError):
-            remote_value.to_knx("100")
+        remote_value = RemoteValue1Count(xknx)
+        self.assertEqual(remote_value.from_knx(DPTArray((0x64, ))), 100)
 
     def test_set(self):
         """Test setting value."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueSceneNumber(
+        remote_value = RemoteValue1Count(
             xknx,
             group_address=GroupAddress("1/2/3"))
-        self.loop.run_until_complete(asyncio.Task(remote_value.set(11)))
+        self.loop.run_until_complete(asyncio.Task(remote_value.set(100)))
         self.assertEqual(xknx.telegrams.qsize(), 1)
         telegram = xknx.telegrams.get_nowait()
         self.assertEqual(
             telegram,
             Telegram(
                 GroupAddress('1/2/3'),
-                payload=DPTArray((0x0A, ))))
-        self.loop.run_until_complete(asyncio.Task(remote_value.set(12)))
+                payload=DPTArray((0x64, ))))
+        self.loop.run_until_complete(asyncio.Task(remote_value.set(101, )))
         self.assertEqual(xknx.telegrams.qsize(), 1)
         telegram = xknx.telegrams.get_nowait()
         self.assertEqual(
             telegram,
             Telegram(
                 GroupAddress('1/2/3'),
-                payload=DPTArray((0x0B, ))))
+                payload=DPTArray((0x65, ))))
 
     def test_process(self):
         """Test process telegram."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueSceneNumber(
+        remote_value = RemoteValue1Count(
             xknx,
             group_address=GroupAddress("1/2/3"))
         telegram = Telegram(
             group_address=GroupAddress("1/2/3"),
-            payload=DPTArray((0x0A, )))
+            payload=DPTArray((0x64, )))
         self.loop.run_until_complete(asyncio.Task(remote_value.process(telegram)))
-        self.assertEqual(remote_value.value, 11)
+        self.assertEqual(remote_value.value, 100)
 
     def test_to_process_error(self):
         """Test process errornous telegram."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueSceneNumber(
+        remote_value = RemoteValue1Count(
             xknx,
             group_address=GroupAddress("1/2/3"))
         with self.assertRaises(CouldNotParseTelegram):
