+++ test/remote_value_tests/remote_value_color_rgb_test.py
@@ -1,15 +1,15 @@
-"""Unit test for RemoteValueDptValue1Ucount objects."""
+"""Unit test for RemoteValueColorRGB objects."""
 import asyncio
 import unittest
 
 from xknx import XKNX
-from xknx.devices import RemoteValueDptValue1Ucount
 from xknx.exceptions import ConversionError, CouldNotParseTelegram
 from xknx.knx import DPTArray, DPTBinary, GroupAddress, Telegram
+from xknx.remote_value import RemoteValueColorRGB
 
 
-class TestRemoteValueDptValue1Ucount(unittest.TestCase):
-    """Test class for RemoteValueDptValue1Ucount objects."""
+class TestRemoteValueColorRGB(unittest.TestCase):
+    """Test class for RemoteValueColorRGB objects."""
 
     def setUp(self):
         """Set up test class."""
@@ -23,63 +23,70 @@ class TestRemoteValueDptValue1Ucount(unittest.TestCase):
     def test_to_knx(self):
         """Test to_knx function with normal operation."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueDptValue1Ucount(xknx)
-        self.assertEqual(remote_value.to_knx(10), DPTArray((0x0A, )))
+        remote_value = RemoteValueColorRGB(xknx)
+        self.assertEqual(remote_value.to_knx((100, 101, 102)), DPTArray((0x64, 0x65, 0x66)))
+        self.assertEqual(remote_value.to_knx([100, 101, 102]), DPTArray((0x64, 0x65, 0x66)))
 
     def test_from_knx(self):
         """Test from_knx function with normal operation."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueDptValue1Ucount(xknx)
-        self.assertEqual(remote_value.from_knx(DPTArray((0x0A, ))), 10)
+        remote_value = RemoteValueColorRGB(xknx)
+        self.assertEqual(remote_value.from_knx(DPTArray((0x64, 0x65, 0x66))), (100, 101, 102))
 
     def test_to_knx_error(self):
         """Test to_knx function with wrong parametern."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueDptValue1Ucount(xknx)
+        remote_value = RemoteValueColorRGB(xknx)
         with self.assertRaises(ConversionError):
-            remote_value.to_knx(256)
+            remote_value.to_knx((100, 101, 102, 103))
         with self.assertRaises(ConversionError):
-            remote_value.to_knx("256")
+            remote_value.to_knx((100, 101, 256))
+        with self.assertRaises(ConversionError):
+            remote_value.to_knx((100, -101, 102))
+        with self.assertRaises(ConversionError):
+            remote_value.to_knx(("100", 101, 102))
+        with self.assertRaises(ConversionError):
+            remote_value.to_knx("100, 101, 102")
 
     def test_set(self):
         """Test setting value."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueDptValue1Ucount(
+        remote_value = RemoteValueColorRGB(
             xknx,
             group_address=GroupAddress("1/2/3"))
-        self.loop.run_until_complete(asyncio.Task(remote_value.set(10)))
+        self.loop.run_until_complete(asyncio.Task(remote_value.set((100, 101, 102))))
         self.assertEqual(xknx.telegrams.qsize(), 1)
         telegram = xknx.telegrams.get_nowait()
         self.assertEqual(
             telegram,
             Telegram(
                 GroupAddress('1/2/3'),
-                payload=DPTArray((0x0A, ))))
-        self.loop.run_until_complete(asyncio.Task(remote_value.set(11)))
+                payload=DPTArray((0x64, 0x65, 0x66))))
+        self.loop.run_until_complete(asyncio.Task(remote_value.set((100, 101, 104))))
         self.assertEqual(xknx.telegrams.qsize(), 1)
         telegram = xknx.telegrams.get_nowait()
         self.assertEqual(
             telegram,
             Telegram(
                 GroupAddress('1/2/3'),
-                payload=DPTArray((0x0B, ))))
+                payload=DPTArray((0x64, 0x65, 0x68))))
 
     def test_process(self):
         """Test process telegram."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueDptValue1Ucount(
+        remote_value = RemoteValueColorRGB(
             xknx,
             group_address=GroupAddress("1/2/3"))
         telegram = Telegram(
             group_address=GroupAddress("1/2/3"),
-            payload=DPTArray((0x0A, )))
+            payload=DPTArray((0x64, 0x65, 0x66)))
         self.loop.run_until_complete(asyncio.Task(remote_value.process(telegram)))
-        self.assertEqual(remote_value.value, 10)
+        self.assertEqual(remote_value.value, (100, 101, 102))
 
     def test_to_process_error(self):
         """Test process errornous telegram."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueDptValue1Ucount(
+        remote_value = RemoteValueColorRGB(
             xknx,
             group_address=GroupAddress("1/2/3"))
         with self.assertRaises(CouldNotParseTelegram):
@@ -90,5 +97,5 @@ class TestRemoteValueDptValue1Ucount(unittest.TestCase):
         with self.assertRaises(CouldNotParseTelegram):
             telegram = Telegram(
                 group_address=GroupAddress("1/2/3"),
-                payload=DPTArray((0x64, 0x65, )))
+                payload=DPTArray((0x64, 0x65, 0x66, 0x67)))
             self.loop.run_until_complete(asyncio.Task(remote_value.process(telegram)))
