+++ test/remote_value_tests/remote_value_color_rgbw_test.py
@@ -3,9 +3,9 @@ import asyncio
 import unittest
 
 from xknx import XKNX
-from xknx.devices import RemoteValueColorRGBW
 from xknx.exceptions import ConversionError, CouldNotParseTelegram
 from xknx.knx import DPTArray, DPTBinary, GroupAddress, Telegram
+from xknx.remote_value import RemoteValueColorRGBW
 
 
 class TestRemoteValueColorRGBW(unittest.TestCase):
