+++ test/remote_value_tests/remote_value_scene_number_test.py
@@ -1,15 +1,15 @@
-"""Unit test for RemoteValueColorRGB objects."""
+"""Unit test for RemoteValueSceneNumber objects."""
 import asyncio
 import unittest
 
 from xknx import XKNX
-from xknx.devices import RemoteValueColorRGB
 from xknx.exceptions import ConversionError, CouldNotParseTelegram
 from xknx.knx import DPTArray, DPTBinary, GroupAddress, Telegram
+from xknx.remote_value import RemoteValueSceneNumber
 
 
-class TestRemoteValueColorRGB(unittest.TestCase):
-    """Test class for RemoteValueColorRGB objects."""
+class TestRemoteValueSceneNumber(unittest.TestCase):
+    """Test class for RemoteValueSceneNumber objects."""
 
     def setUp(self):
         """Set up test class."""
@@ -23,70 +23,63 @@ class TestRemoteValueColorRGB(unittest.TestCase):
     def test_to_knx(self):
         """Test to_knx function with normal operation."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueColorRGB(xknx)
-        self.assertEqual(remote_value.to_knx((100, 101, 102)), DPTArray((0x64, 0x65, 0x66)))
-        self.assertEqual(remote_value.to_knx([100, 101, 102]), DPTArray((0x64, 0x65, 0x66)))
+        remote_value = RemoteValueSceneNumber(xknx)
+        self.assertEqual(remote_value.to_knx(11), DPTArray((0x0A, )))
 
     def test_from_knx(self):
         """Test from_knx function with normal operation."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueColorRGB(xknx)
-        self.assertEqual(remote_value.from_knx(DPTArray((0x64, 0x65, 0x66))), (100, 101, 102))
+        remote_value = RemoteValueSceneNumber(xknx)
+        self.assertEqual(remote_value.from_knx(DPTArray((0x0A, ))), 11)
 
     def test_to_knx_error(self):
         """Test to_knx function with wrong parametern."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueColorRGB(xknx)
+        remote_value = RemoteValueSceneNumber(xknx)
         with self.assertRaises(ConversionError):
-            remote_value.to_knx((100, 101, 102, 103))
+            remote_value.to_knx(100)
         with self.assertRaises(ConversionError):
-            remote_value.to_knx((100, 101, 256))
-        with self.assertRaises(ConversionError):
-            remote_value.to_knx((100, -101, 102))
-        with self.assertRaises(ConversionError):
-            remote_value.to_knx(("100", 101, 102))
-        with self.assertRaises(ConversionError):
-            remote_value.to_knx("100, 101, 102")
+            remote_value.to_knx("100")
 
     def test_set(self):
         """Test setting value."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueColorRGB(
+        remote_value = RemoteValueSceneNumber(
             xknx,
             group_address=GroupAddress("1/2/3"))
-        self.loop.run_until_complete(asyncio.Task(remote_value.set((100, 101, 102))))
+        self.loop.run_until_complete(asyncio.Task(remote_value.set(11)))
         self.assertEqual(xknx.telegrams.qsize(), 1)
         telegram = xknx.telegrams.get_nowait()
         self.assertEqual(
             telegram,
             Telegram(
                 GroupAddress('1/2/3'),
-                payload=DPTArray((0x64, 0x65, 0x66))))
-        self.loop.run_until_complete(asyncio.Task(remote_value.set((100, 101, 104))))
+                payload=DPTArray((0x0A, ))))
+        self.loop.run_until_complete(asyncio.Task(remote_value.set(12)))
         self.assertEqual(xknx.telegrams.qsize(), 1)
         telegram = xknx.telegrams.get_nowait()
         self.assertEqual(
             telegram,
             Telegram(
                 GroupAddress('1/2/3'),
-                payload=DPTArray((0x64, 0x65, 0x68))))
+                payload=DPTArray((0x0B, ))))
 
     def test_process(self):
         """Test process telegram."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueColorRGB(
+        remote_value = RemoteValueSceneNumber(
             xknx,
             group_address=GroupAddress("1/2/3"))
         telegram = Telegram(
             group_address=GroupAddress("1/2/3"),
-            payload=DPTArray((0x64, 0x65, 0x66)))
+            payload=DPTArray((0x0A, )))
         self.loop.run_until_complete(asyncio.Task(remote_value.process(telegram)))
-        self.assertEqual(remote_value.value, (100, 101, 102))
+        self.assertEqual(remote_value.value, 11)
 
     def test_to_process_error(self):
         """Test process errornous telegram."""
         xknx = XKNX(loop=self.loop)
-        remote_value = RemoteValueColorRGB(
+        remote_value = RemoteValueSceneNumber(
             xknx,
             group_address=GroupAddress("1/2/3"))
         with self.assertRaises(CouldNotParseTelegram):
@@ -97,5 +90,5 @@ class TestRemoteValueColorRGB(unittest.TestCase):
         with self.assertRaises(CouldNotParseTelegram):
             telegram = Telegram(
                 group_address=GroupAddress("1/2/3"),
-                payload=DPTArray((0x64, 0x65, 0x66, 0x67)))
+                payload=DPTArray((0x64, 0x65, )))
             self.loop.run_until_complete(asyncio.Task(remote_value.process(telegram)))
