+++ test/remote_value_tests/remote_value_switch_test.py
@@ -3,9 +3,9 @@ import asyncio
 import unittest
 
 from xknx import XKNX
-from xknx.devices import RemoteValueSwitch
 from xknx.exceptions import ConversionError, CouldNotParseTelegram
 from xknx.knx import DPTArray, DPTBinary, GroupAddress, Telegram
+from xknx.remote_value import RemoteValueSwitch
 
 
 class TestRemoteValueSwitch(unittest.TestCase):
