+++ xknx/devices/__init__.py
@@ -12,19 +12,6 @@ from .expose_sensor import ExposeSensor
 from .fan import Fan
 from .light import Light
 from .notification import Notification
-from .remote_value import RemoteValue
-from .remote_value_1count import RemoteValue1Count
-from .remote_value_color_rgb import RemoteValueColorRGB
-from .remote_value_color_rgbw import RemoteValueColorRGBW
-from .remote_value_dpt_2_byte_unsigned import RemoteValueDpt2ByteUnsigned
-from .remote_value_dpt_value_1_ucount import RemoteValueDptValue1Ucount
-from .remote_value_scaling import RemoteValueScaling
-from .remote_value_scene_number import RemoteValueSceneNumber
-from .remote_value_sensor import RemoteValueSensor
-from .remote_value_step import RemoteValueStep
-from .remote_value_switch import RemoteValueSwitch
-from .remote_value_temp import RemoteValueTemp
-from .remote_value_updown import RemoteValueUpDown
 from .scene import Scene
 from .sensor import Sensor
 from .switch import Switch
